update flow sync
